Reposition GitHub ribbon
#245 reports that the GitHub ribbon overlaps the navicon rendering it unclickable.

Analysis of the code shows that there is a top rule of 0px for viewports greater than 1024px, which causes this problem. Analysis of the code also shows that for viewports less than 1024px, a top rule is applied which prevents this problem. Removing this media query and applying the top rule to the first occurrence (on Line 5) solves this problem.