Full support to SassDoc
Configured the '@access' to 'private' in every function so it doesn't
appear on the generated files by SassDoc.

Removed some '@requires' used in wrong way and instead of it, using the
description with markdown syntax for proper highlights.

There were some '@warn' used in wrong way too, those were replaced with
'@throw' and the kind of error was described just as before.

The descriptions and '@types' for settings were added too, so now they
will appear on the generated files by SassDoc. The importance and need
for this approach is related for some '@params' are using those
variables as 'default values'.